+++ r/NEWS.md
@@ -4,6 +4,7 @@
  * `hypervolume()` is significantly faster for more than four dimensions (Andreia P. Guerreiro).
  * `hypervolume()` now handles 1D inputs and provides a clear error for 0D inputs (#58).
  *  Add `GPGame` to [Benchmarks](https://multi-objective.github.io/moocore/r/articles/benchmarks.html).
+ * `is_nondominated()` and `filter_dominated()` use a significantly faster algorithm by Kung et al. for dimensions larger than 3.
 
 
 # moocore 0.2.0